+++ extras/conf_test/avr.cpp
@@ -10,7 +10,7 @@ static_assert(ARDUINOJSON_LITTLE_ENDIAN == 1, "ARDUINOJSON_LITTLE_ENDIAN");
 
 static_assert(ARDUINOJSON_USE_DOUBLE == 1, "ARDUINOJSON_USE_DOUBLE");
 
-static_assert(sizeof(ArduinoJson::detail::VariantSlot) == 8,
+static_assert(sizeof(ArduinoJson::detail::VariantSlot) == 6,
               "sizeof(VariantSlot)");
 
 void setup() {}