+++ src/ArduinoJson/Object/JsonPair.hpp
@@ -16,27 +16,27 @@ class JsonPair {
  public:
   // INTERNAL USE ONLY
   JsonPair(detail::ObjectData::iterator iterator,
-           detail::ResourceManager* resources)
-      : iterator_(iterator), resources_(resources) {}
+           detail::ResourceManager* resources) {
+    if (!iterator.done()) {
+      key_ = iterator->asString();
+      iterator.next(resources);
+      value_ = JsonVariant(iterator.data(), resources);
+    }
+  }
 
   // Returns the key.
   JsonString key() const {
-    if (!iterator_.done())
-      return JsonString(iterator_.key(), iterator_.ownsKey()
-                                             ? JsonString::Copied
-                                             : JsonString::Linked);
-    else
-      return JsonString();
+    return key_;
   }
 
   // Returns the value.
   JsonVariant value() {
-    return JsonVariant(iterator_.data(), resources_);
+    return value_;
   }
 
  private:
-  detail::ObjectData::iterator iterator_;
-  detail::ResourceManager* resources_;
+  JsonString key_;
+  JsonVariant value_;
 };
 
 // A read-only key-value pair.
@@ -44,27 +44,27 @@ class JsonPair {
 class JsonPairConst {
  public:
   JsonPairConst(detail::ObjectData::iterator iterator,
-                const detail::ResourceManager* resources)
-      : iterator_(iterator), resources_(resources) {}
+                const detail::ResourceManager* resources) {
+    if (!iterator.done()) {
+      key_ = iterator->asString();
+      iterator.next(resources);
+      value_ = JsonVariantConst(iterator.data(), resources);
+    }
+  }
 
   // Returns the key.
   JsonString key() const {
-    if (!iterator_.done())
-      return JsonString(iterator_.key(), iterator_.ownsKey()
-                                             ? JsonString::Copied
-                                             : JsonString::Linked);
-    else
-      return JsonString();
+    return key_;
   }
 
   // Returns the value.
   JsonVariantConst value() const {
-    return JsonVariantConst(iterator_.data(), resources_);
+    return value_;
   }
 
  private:
-  detail::ObjectData::iterator iterator_;
-  const detail::ResourceManager* resources_;
+  JsonString key_;
+  JsonVariantConst value_;
 };
 
 ARDUINOJSON_END_PUBLIC_NAMESPACE